// apps/murder-hobo-maker/src/app/models/character/abilities-and-skills/proficiency.model.ts
export class Proficiency {
    _id?: string;
    _name!: string;
    _type?: string;
    _subType?: string;
    _source?: string;
}

/**LIST OF PROFICIENCY TYPES*
 * > armor                  *
 * > language               *
 * > ability check          *
 * > tool                   *
 * > weapon                 */

/**LIST OF ARMOR SUB-TYPES  *
 * > light armor            *
 * > medium armor           *
 * > heavy armor            *
 * > shields                */

/**LIST OF WEAPON SUB-TYPES *
 * > simple                 *
 * > martial                */

/*LIST OF LANGUAGE SUB-TYPES*
 * > common                 *
 * > exotic                 */

/**LIST OF TOOL SUB-TYPES   *
 * > artisan                *
 * > instrument             *
 * > game                   *
 * > other                  */

/**LIST OF ABILITYCHECK SUB-TYPES*
 * > saving throw           *
 * > skill                  */